Switch to c-lzo wrapper library
There have been some obscure issues with the Rasky library.  One
defect was worked around via the "RaskyReader", but there is another
presenting as apparent corruption.  Use the C library instead: there's
not a compelling reason to debug that LZO library, though it was
convenient it was in pure Go.
