+++ .travis.yml
@@ -1,6 +1,14 @@
 language: go
+dist: trusty
 go:
   - 1.10.x
+matrix:
+  include:
+    - env: GOTAGS=
+    - env: GOTAGS=lzo
+      addons:
+        apt:
+          packages: liblzo2-dev
 before_deploy:
   - tar -zcf wal-g.linux-amd64.tar.gz -C cmd/wal-g/ wal-g
 deploy:
@@ -12,3 +20,4 @@ deploy:
   on:
     repo: wal-g/wal-g
     tags: true
+    condition: "$GOTAGS = ''"
