+++ Gopkg.lock
@@ -42,6 +42,12 @@
   revision = "2063d937ea693f6f5c8b213906785ff9e36ba0b6"
 
 [[projects]]
+  name = "github.com/cyberdelia/lzo"
+  packages = ["."]
+  revision = "feb520148d8940294afb1e242dc1d2a7c9c34432"
+  version = "1.0"
+
+[[projects]]
   name = "github.com/go-ini/ini"
   packages = ["."]
   revision = "06f5f3d67269ccec1fe5fe4134ba6e982984f7f5"
@@ -82,12 +88,6 @@
   revision = "c605e284fe17294bda444b34710735b29d1a9d90"
 
 [[projects]]
-  branch = "master"
-  name = "github.com/rasky/go-lzo"
-  packages = ["."]
-  revision = "affec0788321cffe2c68821be1e07e87127b17f0"
-
-[[projects]]
   name = "github.com/ulikunitz/xz"
   packages = [
     "internal/hash",
@@ -114,6 +114,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "77837b5e71581236fefcf4ddba2e417f539413a54849a82e9783c7fd19de8cb8"
+  inputs-digest = "6527682c0e55d3c67828fb5fabb577afabd57ecb4ff9a29b700677df843e1664"
   solver-name = "gps-cdcl"
   solver-version = 1
