+++ Gopkg.toml
@@ -30,10 +30,6 @@
   version = "1.3.4"
 
 [[constraint]]
-  branch = "master"
-  name = "github.com/rasky/go-lzo"
-
-[[constraint]]
   name = "github.com/ulikunitz/xz"
   version = "0.5.4"
