+++ Makefile
@@ -3,9 +3,13 @@ PKG_FILES = $(wildcard *.go)
 
 .PHONY : fmt test install all clean
 
+ifdef GOTAGS
+override GOTAGS := -tags $(GOTAGS)
+endif
+
 test: cmd/wal-g/wal-g
 	go list ./... | grep -v 'vendor/' | xargs go vet
-	go test -v
+	go test $(GOTAGS) -v
 
 fmt: $(CMD_FILES) $(PKG_FILES)
 	gofmt -s -w $(CMD_FILES) $(PKG_FILES)
@@ -21,4 +25,4 @@ clean:
 	(cd cmd/wal-g && go clean)
 
 cmd/wal-g/wal-g: $(CMD_FILES) $(PKG_FILES)
-	(cd cmd/wal-g && go build -ldflags "-X main.BuildDate=`date -u +%Y.%m.%d_%H:%M:%S` -X main.GitRevision=`git rev-parse --short HEAD` -X main.WalgVersion=`git tag -l --points-at HEAD`")
+	(cd cmd/wal-g && go build $(GOTAGS) -ldflags "-X main.BuildDate=`date -u +%Y.%m.%d_%H:%M:%S` -X main.GitRevision=`git rev-parse --short HEAD` -X main.WalgVersion=`git tag -l --points-at HEAD`")
